requirements.txt
langchain
langchain-openai
langchain-text-splitters
langchain-community
langgraph
streamlit
pymupdf
faiss-cpu
python-dotenv
